# perf: faster meta serialisation