- remove dead code
- rewrite meta.as_dict
- don't serialise cache keys

#### sales invoice bundle test

**before**

```
In [9]: %timeit json.dumps(bundle, default=json_handler)
26.9 ms ± 398 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

**after**

```
In [5]: %timeit -n20 json.dumps(bundle, default=json_handler)
7.33 ms ± 122 μs per loop (mean ± std. dev. of 7 runs, 20 loops each)
```